requirements.txt
awesome-slugify==1.6.5
blinker==1.4
certifi==2021.5.30
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
regex==2021.8.21
requests
Unidecode==0.4.21
Werkzeug==2.0.1
WTForms==2.3.3
django-prometheus==2.1.0
